In Chefland, there are N cities (numbered 1 through N). Initially, there are no roads between them. The Mayor of Chefland has ordered engineers to build roads across Chefland in such a way that each city is reachable from all other cities.
For each pair of distinct cities i and j, the cost of building a road between these cities is equal to (i∧j), where ∧ denotes the bitwise AND operation. However, leaders of the cities have a weird condition ― they only allow roads with positive costs to be built from their cities.

Help the Mayor find the minimum cost of building roads in the required way or determine that it is impossible.