Given an array arr of size n. You need to reverse the array.
Input Format:
The first line of input contains a single integer T denoting the number of test cases. Then T test cases follow. Each test case consist of two line input, the 1st line is the size of the array and the 2nd line is the integer array elements.

Output Format:
For each testcase, in a new line, print the reversed array.

Your Task:
This is a function problem. You just need to complete the function reverseArray() that takes arr and n as parameters and reverses arr. The driver code then prints arr.

Expected Time Complexity: O(n).
Expected Auxiliary Space: O(1).

Constraints:
1 <= T <= 100
1 <= n  <= 106
1 <= arri <= 106

Example:
Input:
2
5
1 1 2 2 3
2
4 2
Output:
3 2 2 1 1
2 4

Explanation:
Testcase 1: The array is {1 1 2 2 3}. The reversed array is {3 2 2 1 1}.
Testcase 2: The array is {4 2} and the reversed array is {2 4}.