Given a number n the task is to complete the function which returns an integer denoting the smallest number evenly divisible by each number from 1 to n.
Input:
The first line of input contains an integer T denoting the no of test cases then T test cases follow. Each line contains an integer N.

Output:
For each test case output will be in new line denoting the smallest number evenly divisible by each number from 1 to n.

Constraints:
1<=T<=50
1<=n<=25

Example(To be used only for expected output):
Input:
2
3
6
Output:
6
60

Note:The Input/Ouput format and Example given are used for system's internal purpose, and should be used by a user for Expected Output only. As it is a function problem, hence a user should not read any input from stdin/console. The task is to complete the function specified, and not to write the full code.